refactor: streamline criterion services and remove obsolete components
