+++ src/app/shared/service/Menu/ListItemDetails/ListItemDetailsMenuItemsFunctions.service.ts
@@ -1,4 +1,3 @@
-import { CreateCriterionService } from 'src/app/service/Criterion/CreateCriterionService';
 import { CriteriaByIdSearchService } from 'src/app/service/Search/SearchTypes/CriteriaById/CriteriaByIdSearch.service';
 import { CriteriaResultList } from 'src/app/model/Search/ResultList/CriteriaResultList';
 import { CriteriaSearchService } from 'src/app/service/Search/SearchTypes/Criteria/CriteriaSearch.service';
@@ -9,6 +8,7 @@ import { map, switchMap, take } from 'rxjs';
 import { SearchTermDetailsProviderService } from 'src/app/service/Search/SearchTemDetails/SearchTermDetailsProvider.service';
 import { SearchTermDetailsService } from 'src/app/service/Search/SearchTemDetails/SearchTermDetails.service';
 import { SnackbarMessageService } from 'src/app/service/SnackbarMessage.service';
+import { BuildCriterionService } from 'src/app/service/Criterion/Build/BuildCriterionService';
 
 @Injectable({
   providedIn: 'root',
@@ -17,7 +17,7 @@ export class ListItemDetailsMenuItemsFunctionsService {
   constructor(
     private searchService: CriteriaByIdSearchService,
     private criteriaSearchService: CriteriaSearchService,
-    private criterionService: CreateCriterionService,
+    private criterionService: BuildCriterionService,
     private searchTermDetailsService: SearchTermDetailsService,
     private feasibilityQueryProviderHub: FeasibilityQueryProviderHub,
     private searchTermDetailsProviderService: SearchTermDetailsProviderService,
@@ -36,7 +36,7 @@ export class ListItemDetailsMenuItemsFunctionsService {
 
   public addToStage(id: string) {
     this.criterionService
-      .createCriteriaFromHashes([id], false)
+      .buildCriteriaFromHashes([id])
       .pipe(
         map((criteria: Criterion[]) => {
           this.feasibilityQueryProviderHub.addCriteriaToCriterionProvider(criteria);
